Implement Backend and Socket Client Example
Our Backend acts as a "proxy" between the mobile app and the robot. The app communicates with the backend via Websockets, the backend then sends a http requests to the robot to start the driving process and then continuosly polls for status updates, which it relays via a Websocket to the client.